feat: Trying to complete resolving getting crashes

+++ shared/common.h
@@ -7,8 +7,8 @@ extern "C"
     struct StubConfig
     {
         char signature[8];
-        unsigned int packed_data_rva;
-        unsigned int packed_data_size;
-        unsigned int original_data_size;
+        unsigned long packed_data_rva;
+        unsigned long packed_data_size;
+        unsigned long original_data_size;
     } volatile stub_config = { {'H', 'A', 'X', 'O', 'S', 'T', 'U', 'B'} };
 }
